requirements.txt
fastapi==0.116.0
uvicorn==0.25.0
python-multipart==0.0.6
pdfplumber==0.10.1
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==9.5.0   # downgrade from 10.0.0
openpyxl==3.1.2
google-generativeai==0.8.3
python-dotenv==1.0.0
